refactor upsells dashboard
